Restructured files with Destiny
