+++ client/src/index/App/shared/Forum.js
@@ -2,11 +2,11 @@ import { useEffect, React } from "react";
 import Container from "@material-ui/core/Container";
 import Grid from "@material-ui/core/Grid";
 import { Route, Switch, useParams } from "react-router-dom";
-import Menu from "./Menu";
-import Announcements from "./Contents/Announcements";
-import Assignments from "./Contents/Assignments";
-import Files from "./Contents/Files";
-import QnA from "./Contents/QnA";
+import Menu from "../Main/Forum/Menu";
+import Announcements from "../Main/Forum/Announcements";
+import Assignments from "../Main/Forum/Assignments";
+import Files from "../Main/Forum/Files";
+import QnA from "../Main/Forum/QnA";
 import Box from "@material-ui/core/Box";
 import { makeStyles } from "@material-ui/core/styles";
 import KeyboardArrowUpIcon from "@material-ui/icons/KeyboardArrowUp";
